+++ pkg/local_object_storage/shard/delete.go
@@ -83,9 +83,12 @@ func (s *Shard) delete(prm DeletePrm) (DeleteRes, error) {
 	s.decObjectCounterBy(physical, res.RawObjectsRemoved())
 	s.decObjectCounterBy(logical, res.AvailableObjectsRemoved())
 	for i := range prm.addr {
-		removedPayload := res.RemovedObjectSizes()[i]
+		removedPayload := res.RemovedPhysicalObjectSizes()[i]
 		totalRemovedPayload += removedPayload
-		s.addToContainerSize(prm.addr[i].Container().EncodeToString(), -int64(removedPayload))
+		logicalRemovedPayload := res.RemovedLogicalObjectSizes()[i]
+		if logicalRemovedPayload > 0 {
+			s.addToContainerSize(prm.addr[i].Container().EncodeToString(), -int64(logicalRemovedPayload))
+		}
 	}
 	s.addToPayloadSize(-int64(totalRemovedPayload))